+++ src/main/java/com/appcenter/marketplace/domain/tempMarket/service/TempMarketService.java
@@ -6,5 +6,5 @@ import com.appcenter.marketplace.domain.tempMarket.dto.res.TempMarketRes;
 public interface TempMarketService {
     TempMarketPageRes<TempMarketRes> getMarketList(Long memberId, Long marketId, Integer size, String category);
 
-    TempMarketPageRes<TempMarketRes> getUpcomingNearMarketList(Long memberId, Long marketId, Long cheerCount, Integer size);
+    TempMarketPageRes<TempMarketRes> getUpcomingNearMarketList(Long memberId, Long marketId, Integer cheerCount, Integer size);
 }
